fix: Added KeyError exception catching
When trying to retrieve streaming data in a video, a KeyError :
'streamingData' exception will often occur, for example, when trying to
download a video with an age restriction.

However, when downloading
unrestricted videos, this error also often occurs.

Added a decorator
that restarts the function that downloads the video up to 10 times,
which is enough to download videos for which there are no restrictions

Signed-off-by: rin-gil <91410075+rin-gil@users.noreply.github.com>